+++ doc/whats-new.rst
@@ -36,7 +36,8 @@ Deprecations
 
 Bug fixes
 ~~~~~~~~~
-
+- Fixed :py:meth:`Dataset.transpose` to raise a more informative error. (:issue:`6502`, :pull:`7120`)
+  By `Patrick Naylor <https://github.com/patrick-naylor>`_
 
 Documentation
 ~~~~~~~~~~~~~